Support embed record (temporary)
